+++ index.d.ts
@@ -67,11 +67,6 @@ export class LoginClient {
   constructor(settings: Settings, logger?: unknown);
 
   /**
-   * @deprecated
-   */
-  getUserData(): Record<string, unknown>;
-
-  /**
    * @description Gets the user's profile data and returns it if it exists. Should be called after {@link userSessionExists} or it will be empty.
    * @return {Object} The user identity which contains a userData object.
    */