+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This is the changelog for [Authress Login](readme.md).
 * Add automatic retries to network connection issues.
 * Add AuthUserId cookie available in all requests to replace the `user` cookie.
 * Avoid unnecessary CORS warnings when using a cross domain application
+* Improve logout redirect url default location
 
 ## 1.4 ##
 * Include the extensionClient as embeddable client for OAuth extension login.