+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log
 This is the changelog for [Authress Login](readme.md).
 
+## 2.2 ##
+* Automatically retry on network connection issues.
+
 ## 2.1 ##
 * Remove `connectionId` and `tenantLookupIdentifier` requirements from the authentication call so that the user can be directed to the Authress Hosted login when necessary: https://authress.io/app/#/settings?focus=branding
 * Enable the methods to have optional inputs parameters when not required